requirements.txt
os
random
asyncio
requests
gradio
python-dotenv
uvicorn
fastapi
pydantic
langchain
langchain_chroma
langchain_openai
langchain_huggingface
langchain_core
langchain_community
langchain_text_splitters
PyMuPDF
